+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ _Date TBD_
 * Made ICS Source required for 2015 w/ICS on upload/confirm/edit
 * Test tools are required for 170.315 (b)(8)
 * GAP is required for 170.314 (b)(5)(B)
+* Add warning messages to uploaded listings and surveillance regarding invalid characters found.
 
 ---